When reading data out of a field you can now supply a default value that gets returned if the field does not exist (instead of having an error thrown)
Added a static "null" object with type g_null and no fields defined
Added "Exists()" which returns false if the object it's called on is not null

The null object gets returned if you look up a field that does not exist (instead of having an error thrown)
(this means you don't need to check for null after every step in a chain. You can just do mygonobject["foo"]["bar"]["fab"]["amount"].Exists(); It'll return false if any one of those fields returns the null object

Sadly this is at the expense of good error reporting when you access a missing field. I'll figure out how to add that back on later